requirements.txt
pandas==2.2.3
python-dotenv==1.0.1
openai==1.50.2
matplotlib==3.9.2
seaborn==0.13.2
wordllama==0.2.8.post0